Removed file fragmentation
