+++ ld_script.txt
@@ -47,8 +47,6 @@ SECTIONS
         src/rom_1332C.o(.text);
         asm/rom_11B9C.o(.text1333C);
         src/rom_18784.o(.text);
-        asm/rom_11B9C.o(.text18A4C);
-        src/rom_18A4C.o(.text);
         asm/rom_11B9C.o(.text18B50);
         src/rom_19B64.o(.text);
         src/rom_19D04.o(.text);
